product/requirements.txt
django==3.1.5
python-dotenv==0.15.0
pillow==8.1.0
djangorestframework.3.12.2
celery==5.0.5
redis==3.5.3
pytest==6.2.2
pytest-django==4.1.0
rich==9.12.2
